Let child CUDAFuture wait for parent CUDAFuture's CUDAEvents
If the child cannot extract tensors from returned IValue, the
current child CUDAFuture won't wait for anything. In this case,
if the `wait()` wasn't called on the parent Future, streams are
not synchronized, and it is possible that parent Future's CUDA
ops have not been added to streams yet.

This commit adds a `markCompletedWithDataPtrs()` to `ivalue::Future`,
and RPC uses this API to pass Message tensor dataPtrs to the
`PyObject` Future when marking it as completed.

ghstack-source-id: 2d60fbf722d3d34e5b0bb775e1b7bf22c9582c6b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51820