[quant] Quantizable MultiheadAttention
- Adds the `torch.nn.quantizable.MultiheadAttention`

The quantizable version can serve as a fully equivalent to `torch.nn.MultiheadAttention` module.
The main difference is that it allows for linear units observation after the `prepare` step in the quantization flow.

Note: The `from_observed` method (called during the `convert`) removes the `bias_k` and `bias_v` parameters, and resets them as attributes.
This is done to avoid an error of assigning a quantized tensor to the `torch.nn.Parameter`.

Test Plan:

```
python test/test_quantization.py TestQuantizedOps.test_custom_module_multi_head_attention
```

ghstack-source-id: b451d0f6cf107ad16bcfcb2b49ac999d9c04c7fb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49866