[quant][refactor] Factor out MHA code dup from nn and nn.quantizable
Quantizable MHA shares some code with its FP equivalent.
This factors out common parts.

Test Plan:

Explicit testing is not available, as there is no functional tests for the nn.MHA.
However, the functional equivalence of this PR was tested across commits by comparing the numerics of the pre-change and post-change MHA. Similar was done for the quantizable counterpart.

ghstack-source-id: 065da6ed2b4b9a9896945c18af7921e3750dfc81
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51169