[te][llvm] Generate arithmetic vs logical right shift as appropriate
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51749

Following in the mode of C++, we probably want to distinguish when
it's appropriate to do arithmetic vs. logical right shift.

> For negative a, the value of a >> b is implementation-defined (in most
> implementations, this performs arithmetic right shift, so that the result
> remains negative).

If you look at what clang does, if `a` is unsigned, a logical shift is
generated; if signed, an arithmetic shift.  Let's do the same here.  This turns
out to be useful for, e.g., implementing transcendental function
approximations.
ghstack-source-id: 121366317

Differential Revision: [D26245856](https://our.internmc.facebook.com/intern/diff/D26245856/)